LayoutTests/fast/dom/SelectorAPI/dumpNodeList-almost-strict-expected.txt

Line 1
Line 2
Line 3

line 4
line 5

Document.querySelectorAll

[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 4
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 5
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
[object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], [object HTMLDivElement], length: 4

Element.querySelectorAll

length: 0
[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLParagraphElement], length: 2
[object HTMLDivElement], [object HTMLParagraphElement], length: 2
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12

DocumentFragment.querySelectorAll

[object HTMLDivElement], length: 1
[object HTMLDivElement], length: 1
length: 0
[object HTMLDivElement], [object HTMLDivElement], length: 2
[object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 3
[object HTMLDivElement], [object HTMLDivElement], [object HTMLParagraphElement], length: 3
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12

Document.querySelector

[object HTMLDivElement]
TypeError: Result of expression 'node.querySelector(selectorString)' [null] is not an object.
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12

Element.querySelector

TypeError: Result of expression 'node.querySelector(selectorString)' [null] is not an object.
[object HTMLDivElement]
TypeError: Result of expression 'node.querySelector(selectorString)' [null] is not an object.
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12

DocumentFragment.querySelector

[object HTMLDivElement]
[object HTMLDivElement]
TypeError: Result of expression 'node.querySelector(selectorString)' [null] is not an object.
[object HTMLDivElement]
[object HTMLDivElement]
[object HTMLDivElement]
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
Error: SYNTAX_ERR: DOM Exception 12
